feat: Update media-go-pkg to v0.3.0
The media-service module now requires version 0.3.0 of the media-go-pkg package instead of v0.1.8, which was previously required. This change also updates the go.sum file with the new hash for v0.3.0 of media-go-pkg package and its corresponding go.mod file entry has been updated as well in go.mod file.