refactor: Update models.go to include functions for converting arrays of movies and TV shows
This commit adds two new functions to the models.go file that allow for conversion of arrays of movies and TV shows into their respective response types. The `toMoviesResponse` function converts an array of `tmdb.Movie` structs into an array of `movieResponse` structs, while the `toTVShowsResponse` function does the same for an array of `tmdb.TVShow` structs. These functions will be useful in future development when returning multiple movies or TV shows in a single API call.