feat(media): Add functions to get media rating, media info, episode and tv show
This commit adds new functions to the MediaRepository struct that allow getting the average rating and number of ratings for a media given its ID, as well as retrieving information about a specific media (either movie or TV show), an episode or file info for either a movie or an episode. The code has also been refactored to use more efficient queries.