feat(media_info): Add functions to retrieve media, movie, tv show and episode info
This commit adds new functions to the MediaData struct that allow retrieval of media information from TMDB. Specifically, it includes GetMediaByID(), GetMovieInfo(), GetTvShowInfo() and GetEpisodeInfo(). These functions take in a TMDB ID or TV ID along with other parameters such as season number and episode number (for episodes) and return the corresponding information.