feat(models): add Present field to responses
This commit adds a new boolean field "Present" to the movie, TV show and TV episode response structs. The value of this field is set by passing a boolean argument to the corresponding conversion functions. Additionally, the conversion functions for movies, TV shows and TV episodes now accept an array of booleans that indicate whether each item in the input slice was present or not. This allows us to include information about whether a particular movie, TV show or episode was available when generating responses.