feat: Update media-go-pkg to v0.5.4 and add TV show search by actor
This commit updates the media-go-pkg dependency from v0.5.3 to v0.5.4, which includes bug fixes and performance improvements.

Additionally, this commit adds a new feature that allows users to search for TV shows by actor ID in the /discover/tv/actor endpoint.

The getTvShowsByActor function was added to internal/controllers/discover.go, while GetShowsByActor was added to internal/features/media_discover.go as an implementation of the new feature.