feat: Add functions to get followed movie and TV show releases by month
Two new functions were added to the MediaRepository struct. The GetFollowedMovieReleases function returns a list of movies that a user is following with release dates in a given month, while the GetFollowedTvShowReleases function does the same for TV shows. These functions use SQL queries to retrieve data from the database based on user ID and release date.