feat: Add month query parameter to calendar endpoints
This commit adds a new required query parameter "month" to the calendar endpoints. The value of this parameter must be an integer between 1 and 12, representing the month for which the user wants to retrieve their movie or TV show calendar. If the parameter is not provided or has an invalid value, an error response will be returned.

Additionally, this commit updates the getMoviesCalendar and getTvShowsCalendar functions in calendar.go to pass the month parameter to their respective service methods. It also introduces a new tvReleasesResults struct and a corresponding toTVReleasesResult function in models.go that combines TV episode and TV show responses into one JSON object.