feat: Add TMDB package to media repository
This commit adds the TMDB package to the media repository. It also includes changes to retrieve movie and TV show ratings, as well as file information for movies and episodes. Additionally, it adds functions to check if a movie, TV show, episode, or file is present in the database. Finally, it updates functions that return lists of movies and TV shows to only include available ones with non-null media files.