feat: Add object storage support for media files
This commit adds support for uploading and deleting media files to an object storage service. The `MediaFile` struct now includes an `objectStorage` field, which is used to upload and delete media files instead of using the local file system. The `DeleteMediaFile` function has been updated to use the new `objectStorage` field instead of removing folders from the local file system.